move "StyledComponentsRegistry" to Styles Folder
