+++ styles/styledComponentRegistry.jsx
